feat: Update psi.md (#631)
* feat: Update psi.md

Their study the problem of private set-intersection in the multi-party setting and design two protocols with the following improvements compared to prior work. First, Their protocols are designed in the so-called star network topology, where a designated party communicates with everyone else, and take a new approach of leveraging the 2PC protocol. This approach minimizes the usage of a broadcast channel, where our semi-honest protocol does not make any use of such a channel and all communication is via point-to-point channels.In addition, the communication complexity of our protocols scales with the number of parties.

* style: add 2 additional space