Update survey.md (#567)
* Update survey.md

* Update survey.md

**Type of change**

- [x] Add new papers (Please tell us why you think this paper is awesome!)
- [ ] Fix the category of an existing paper/papers (Please tell us the reasons)
- [ ] Add a new tool/primitive/application with a new markdown page (Thank you! Also, please tell us more about this awesome thing!)

**Description**

> Here is an updated survey paper on federated learning in the direction of **Security** and **Personalization** that can help us systematically learn about backdoor attacks and personalization in federated learning. 

Add a new **PMLR20** paper:  How to Backdoor Federated Learning. This is the first paper detailing the process of backdoor attacks in federated learning.

Add a new **TNNLS22** paper: Towards Personalized Federated Learning. This paper highlights the key ideas and challenges of personalized federated learning, with a unique classification and future outlook.

@jamie-cui